# By m9710797 (25) and others
# Via Maarten ter Huurne
* 'master' of ssh://git.code.sf.net/p/openmsx/openmsx: (40 commits)
  Renamed getSDLWindowHandle() to getWindowHandle()
  Always init video system before getting window handle
  Fixed check of return code from SDL_GetWMInfo()
  Removed unnecessary include of "sdlwin32.hh"
  Use native Windows function to get thread ID in MidiInWindows
  Better use 32x32 icons on Windows, it seems.
  Load a hi-res icon, if available.
  Revert to defaults when no SDL video surface is available yet.
  Fixed crash (only) in debug build
  Revert "removed workaround for gcc-4.6"
  disable -Wzero-as-null-pointer-constant warning, see comments in patch
  removed workaround for gcc-4.6
  Remove gcc-3.4 workaround
  Reduce size of LUT for rapidsax
  workaround (harmless) visual studio compiler warnings
  Work around visual studio compiler error
  [13/13] map->vector: breakpoints
  [12/13] map->vector: android pressed keys
  [11/13] map->vector: MSXMixer SoundDeviceInfo
  [10/13] map->vector: small cleanup in HotKey
  ...

+++ src/fdc/DirAsDSK.cc
@@ -639,17 +639,17 @@ void DirAsDSK::setMSXTimeStamp(DirIndex dirIndex, FileOperations::Stat& fst)
 }
 
 // Used to add 'regular' files before 'derived' files. E.g. when editing a file
-// in a host editor, you often get a backup/swap fileis like this:
+// in a host editor, you often get backup/swap files like this:
 //   myfile.txt  myfile.txt~  .myfile.txt.swp
 // Currently the 1st and 2nd are mapped to the same MSX filename. If more
 // host files map to the same MSX file then (currently) one of the two is
 // ignored. Which one is ignored depends on the order in which they are added
 // to the virtual disk. This routine/heuristic tries to add 'regular' files
 // before derived files.
-static int weight(const string& hostName)
+static size_t weight(const string& hostName)
 {
 	// TODO this weight function can most likely be improved
-	int result = 0;
+	size_t result = 0;
 	string_ref file, ext;
 	StringOp::splitOnLast(hostName, '.', file, ext);
 	// too many '.' characters
